requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
mariadb==1.1.12
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
